Dissect AzureLoginHelper.ts into multiple files
