+++ src/extensionVariables.ts
@@ -5,10 +5,12 @@
 
 import { ExtensionContext } from "vscode";
 import { IAzExtOutputChannel, IExperimentationServiceAdapter } from "vscode-azureextensionui";
+import { AzureAccountLoginHelper } from "./login/AzureLoginHelper";
 import { UriEventHandler } from "./login/exchangeCodeForToken";
 
 export namespace ext {
     export let context: ExtensionContext;
+    export let loginHelper: AzureAccountLoginHelper;
     export let outputChannel: IAzExtOutputChannel;
     export let uriEventHandler: UriEventHandler;
     export let experimentationService: IExperimentationServiceAdapter;
