+++ src/login/adal/getUserCode.ts
