+++ src/login/getCurrentTarget.ts
@@ -1,0 +1,19 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { ConfigurationTarget } from "vscode";
+
+export function getCurrentTarget(config: { key: string; defaultValue?: unknown; globalValue?: unknown; workspaceValue?: unknown, workspaceFolderValue?: unknown } | undefined): ConfigurationTarget {
+	if (config) {
+		if (config.workspaceFolderValue) {
+			return ConfigurationTarget.WorkspaceFolder;
+		} else if (config.workspaceValue) {
+			return ConfigurationTarget.Workspace;
+		} else if (config.globalValue) {
+			return ConfigurationTarget.Global;
+		}
+	}
+	return ConfigurationTarget.Global;
+}
